HW3: add checking for user_id(as some hackers could change meal id and remove different users data)

+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcUserMealRepositoryImpl.java
@@ -83,12 +83,12 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
 
     @Override
     public boolean delete(int id, int userId) {
-        return jdbcTemplate.update("DELETE FROM meals WHERE id=?", id) != 0;
+        return jdbcTemplate.update("DELETE FROM meals WHERE id=? AND user_id=?", id, userId) != 0;
     }
 
     @Override
     public UserMeal get(int id, int userId) {
-        List<UserMeal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=? ORDER BY dateTime DESC", ROW_MAPPER, id);
+        List<UserMeal> meals = jdbcTemplate.query("SELECT * FROM meals WHERE id=? AND user_id=? ORDER BY dateTime DESC", ROW_MAPPER, id, userId);
         return DataAccessUtils.singleResult(meals);
     }
 
@@ -99,6 +99,6 @@ public class JdbcUserMealRepositoryImpl implements UserMealRepository {
 
     @Override
     public List<UserMeal> getBetween(LocalDateTime startDate, LocalDateTime endDate, int userId) {
-        return jdbcTemplate.query("SELECT * FROM meals WHERE dateTime > ? AND dateTime < ? ORDER BY dateTime DESC", ROW_MAPPER, Timestamp.valueOf(startDate), Timestamp.valueOf(endDate));
+        return jdbcTemplate.query("SELECT * FROM meals WHERE dateTime > ? AND dateTime < ? AND user_id=? ORDER BY dateTime DESC", ROW_MAPPER, Timestamp.valueOf(startDate), Timestamp.valueOf(endDate), userId);
     }
 }
